Enigma is parent class to new decrypt and encrypt classes, move tests to reflect this.
