+++ encrypt.rb
@@ -1,11 +1,11 @@
-require_relative './lib/enigma'
+require_relative './lib/encrypt'
 
 ARGV
 plain_message_path = File.open("files/#{ARGV[0]}")
 encrypted_message_file = File.open("files/#{ARGV[1]}", "w" )
 message = plain_message_path.read.chomp
 
-enigma = Enigma.new
+enigma = Encrypt.new
 info = enigma.encrypt("#{message}")
 encrypted_message_file.write("#{info[:encryption]}")
